Factor out ScaffoldOperationFactory
